Adapted chat system to use the University of Bath's linux server. This means we can now run the server on the university's server and then connect to the chat server from multiple computers.
It works using SSH port forwarding, it will sign into my personal university account on the linux server and port forward the localhost port, 14002, to the remote port 34751. (Private key is not included in this github repo for security reasons).

I added a feature for the client to give a name so when the server broadcasts messages it says who they're from.