Installed flatlaf package to modernise the GUI with a new Swing look and feel.
Allowed safe quiting of the program. When closing the program it will inform the server, cancel the Swing worker listening for messages, disconnect the session and delete the port forwarding.

Added useful messages from the server to be sent to the clients when you connect/disconnect.

General cleanup of code, adding comments and removing legacy code.